+++ giswater_admin/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Schema upgrade failing with `must be owner of function` during `reload_fct_ftrg` when update patches redefined functions as the installer superuser.
 ## [0.3.0] - 2026-06-19
 
 ### Added